Move some classes out from PdlLeesahConsumer.kt
